Remove useless status
Removed "done" status which was meant to be used when a show
has finished airing. It was meant to be detected
from self.info (TV Maze API data), but that's only available
when updating a show and which in turn makes it useless (because
it's supposed to work "offline" from the cached data as well).